requirements.txt
beautifulsoup4==4.3.2
gnureadline==6.3.3
ipython==2.3.1
python-firebase==1.2
requests==2.3.0
wikipedia==1.3.1
retrying==1.3.2
